Add files via upload
This is notification system which gives desktop notifications for drinking water and taking a break. The system also provides a different benefit of drinking water and take a break quote along with the notification. You can set the time according to your schedule. This program gives notification to drink water every hour and reminds you to take break every 3 hours.